docs: API reference, tutorial restructure, admonition fixes
- book.yml: add auto-generated API reference (api_docs; curated user-facing
  packages, MPDetector excluded); drop the dead `usage_guide` TOC entry
  (content folded into FAQs); point the Tutorials TOC at the renamed notebooks.
- Tutorials: rename to descriptive names (Detecting_Images, Detecting_Videos,
  Plotting, Analysis, Performance); remove the redundant 05_detectorv2 — its
  valence/arousal demo is folded into Detecting_Images and its interactive
  mesh is already covered by Plotting §3.6; byline/Colab-cell cleanup.
- pages: convert MyST ```{note}``` admonitions to Material `!!! note`
  (models.md, faqs.md).
- _rendered/: regenerate committed cached bodies for the renamed paths.
- requirements-dev.txt: add marimo-book[api] for the API-reference build.
