+++ docs/marimo/content/02_detector_vids.py
@@ -15,9 +15,6 @@ def _():
 def _(mo):
     mo.md(r"""
     # 2. Detecting facial expressions from videos
-
-    *Written by Jin Hyun Cheong and Eshin Jolly*
-
     In this tutorial we'll explore how to use the `Detector` class to process video files. You can try it out interactively in Google Collab: [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/cosanlab/py-feat/blob/master/notebooks/content/03_detector_vids.ipynb)
     """)
     return
@@ -45,7 +42,7 @@ def _():
     from feat import Detector
 
     detector = Detector()
-    return (detector,)
+    return 
 
 
 @app.cell(hide_code=True)
